train_scripts/requirements.txt
torchvision==0.12.0
datasets==2.5.2
transformers==4.17.0
